[*] refactoring

+++ src/main/java/dao/InstructorDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Set;
 
 public interface InstructorDAO {
-    Instructor create(String firstName, String secondName, byte age);
+    Instructor create(String firstName, String secondName, int age);
     List<Instructor> getAll();
     List<Instructor> findByFirstName(String firstName);
 }
